requirements.txt
dkimpy==1.1.8
requests==2.32.4
pydantic[email]==2.11.5
pydantic-settings==2.9.1
pysocks==1.7.1
